fix: use i64 for shards/ranges
Using int is problematic because it's platform-dependent. On 32-bit
platform this may cause issues if SubjectShards > 2,147,483,647. This
will make the number go negative and ranges might not match.

+++ eppoclient/configurationstore.go
@@ -22,7 +22,7 @@ type Allocation struct {
 type experimentConfiguration struct {
 	Name          string                `json:"name"`
 	Enabled       bool                  `json:"enabled"`
-	SubjectShards int                   `json:"subjectShards"`
+	SubjectShards int64                 `json:"subjectShards"`
 	Rules         []rule                `json:"rules"`
 	Overrides     map[string]Value      `json:"typedOverrides"`
 	Allocations   map[string]Allocation `json:"allocations"`
